// api/modules/opportunity/clubs_v2/types.ts
export const opportunityClubV2Include = {
  theme: true,
  themeVariant: true,
} as const;